backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
boto3==1.34.69
pandas==2.2.1
pyarrow==15.0.2
python-dotenv==1.0.1
pydantic==2.7.0
python-jose[cryptography]==3.5.0
motor==3.7.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
numpy==1.26.4